Extract TOML extension to own module
