+++ libimagstore/src/lib.rs
@@ -52,4 +52,5 @@ pub mod hook;
 pub mod store;
 mod configuration;
 mod file_abstraction;
+mod toml_ext;
